requirements.txt
apache-airflow==2.7.2
pandas==2.0.3